이전과제 코드정리
