+++ Homework/Kyusik/RxStudy/app/src/main/res/layout/activity_observer_pattern.xml
@@ -1,0 +1,125 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:padding="30dp"
+        tools:context=".ObserverPatternActivity">
+
+        <TextView
+            android:id="@+id/tv_opacity_title"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/first_week_observer_pattern_opacity_title"
+            android:textSize="13sp"
+            app:layout_constraintBottom_toBottomOf="@id/et_percent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="@id/et_percent" />
+
+        <com.rxstudy.customui.PercentEditText
+            android:id="@+id/et_percent"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="30dp"
+            android:hint="@string/first_week_observer_pattern_opacity_hint"
+            android:inputType="number"
+            android:textSize="13sp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toEndOf="@id/tv_opacity_title"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <TextView
+            android:id="@+id/tv_opacity_seek_bar_title"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="30dp"
+            android:text="@string/first_week_observer_pattern_opacity_seek_bar_title"
+            android:textSize="13sp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/et_percent" />
+
+        <com.rxstudy.customui.PercentSeekBar
+            android:id="@+id/sb_percent"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            app:layout_constraintBottom_toBottomOf="@id/tv_opacity_seek_bar_title"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toEndOf="@id/tv_opacity_seek_bar_title"
+            app:layout_constraintTop_toTopOf="@id/tv_opacity_seek_bar_title" />
+
+        <com.rxstudy.customui.PercentImageView
+            android:id="@+id/img_opacity_result"
+            android:layout_width="30dp"
+            android:layout_height="30dp"
+            android:layout_marginTop="30dp"
+            android:background="@color/black"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/sb_percent" />
+
+        <com.rxstudy.customui.PercentTextView
+            android:id="@+id/tv_percent"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="30dp"
+            android:textSize="13sp"
+            app:layout_constraintBottom_toBottomOf="@id/img_opacity_result"
+            app:layout_constraintStart_toEndOf="@id/img_opacity_result"
+            app:layout_constraintTop_toTopOf="@id/img_opacity_result" />
+
+
+        <TextView
+            android:id="@+id/tv_graph_title"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="30dp"
+            android:text="@string/graph_title"
+            android:textSize="13sp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/img_opacity_result" />
+
+        <com.rxstudy.customui.GraphEditText
+            android:id="@+id/et_graph"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="30dp"
+            android:hint="@string/graph_hint"
+            android:inputType="number"
+            android:textSize="13sp"
+            app:layout_constraintBottom_toBottomOf="@+id/tv_graph_title"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toEndOf="@id/tv_graph_title"
+            app:layout_constraintTop_toTopOf="@+id/tv_graph_title" />
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/bar_container"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="30dp"
+            android:background="@drawable/graph_bar_container"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/et_graph">
+
+            <com.rxstudy.customui.GraphImageView
+                android:id="@+id/img_graph"
+                android:layout_width="match_parent"
+                android:layout_height="10dp"
+                android:layout_marginVertical="10dp"
+                android:background="@drawable/graph_bar_mango"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent" />
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
+        <com.rxstudy.customui.GraphTextView
+            android:id="@+id/tv_graph"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textSize="13sp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/bar_container" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
