+++ Homework/Kyusik/RxStudy/app/src/main/res/layout/activity_subject_example.xml
@@ -6,26 +6,32 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:padding="20dp"
         android:orientation="vertical"
-        tools:context=".MainActivity">
+        android:padding="20dp"
+        tools:context=".SubjectExampleActivity">
 
         <Button
-            android:id="@+id/bt_observer_pattern"
+            android:id="@+id/bt_publish_subject"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="Observer Pattern" />
+            android:text="publish subject" />
 
         <Button
-            android:id="@+id/bt_stream"
+            android:id="@+id/bt_behavior_subject"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="Stream Example" />
+            android:text="Behavior Subject " />
 
         <Button
-            android:id="@+id/bt_subject"
+            android:id="@+id/bt_async_subject"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="Subject Example" />
+            android:text="Async Subject " />
+
+        <Button
+            android:id="@+id/bt_replay_subject"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Replay Subject" />
     </LinearLayout>
 </layout>
